requirements.txt
# Phoenix Core 核心依赖
# 适用于 Python 3.10+

# Discord Bot
discord.py>=2.3.0

# Web Framework (FastAPI + Uvicorn)
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# HTTP Client
requests>=2.31.0
aiohttp>=3.9.0

# Database
sqlite3  # Python 内置

# Office Document Support
openpyxl>=3.1.0  # Excel
python-docx>=1.0.0  # Word
python-pptx>=1.0.0  # PowerPoint
Pillow>=10.0.0  # 图片处理

# LLM API
dashscope>=1.14.0  # 通义千问

# Utilities
python-dotenv>=1.0.0  # 环境变量
pyyaml>=6.0  # YAML 支持
jinja2>=3.1.0  # 模板引擎

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-benchmark>=4.0.0

# Code Quality (开发依赖)
flake8>=6.1.0
black>=23.0.0
mypy>=1.5.0
